requirements.txt
pytest==9.0.2
pytest-bdd==8.1.0
requests==2.32.5
pathlib==1.0.1
python-dotenv==1.0.1
allure-pytest==2.13.5